Cancel Slurm job solution (#8)
* added an issue solution

* fixed the obtaining of a deleting-job owner name, changed jobId to long type, fixed squeue-command execution bug, some refactoring SlurmJobProviderTests

* changes according to review comments

* Merge branch 'slurm_develop' into #11_slurm_job_cancelation, refactored shouldMakeValidEnvVariables test

+++ src/main/java/com/epam/grid/engine/entity/JobFilter.java
@@ -37,7 +37,7 @@ public class JobFilter {
     /**
      * List of job IDs.
      */
-    private List<Integer> ids;
+    private List<Long> ids;
     /**
      * State of job.
      */
